Drop, CastToUI for inventory, Duplicatatecs in PrefabHashes

+++ Assets/Crimson.Core/Components/SpawnPrefabData.cs
@@ -5,7 +5,7 @@ namespace Assets.Crimson.Core.Components
 {
     public struct SpawnPrefabData : IBufferElementData
     {
-        public ushort ID;
+        public int ID;
         public Vector3 Position;
         public Quaternion Rotation;
     }
